Fix to get plugin working with Grails 2.2.
Groovy 2, which comes bundled with Grails 2.2, no longer has the
`GroovyShell.evaluate(InputStream)` method. As this was being used by
`FixtureFileLoader` the plugin would not compile.

We use a `Reader` instead of an `InputStream`, specifying the file
encoding based on a new `plugin.fixtures.file.encoding` configuration
option. The default is UTF-8.